requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
Django==3.2.6
django-admin-inline-paginator==0.2
django-extra-views==0.14.0
django-kronos==1.0
filelock==3.0.12
idna==3.2
jsonfield==3.1.0
Naked==0.1.31
pkg_resources==0.0.0
platformdirs==2.2.0
pycodestyle==2.7.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
virtualenv==20.7.0
